Crate prisma_fmt_build

Functions

code_actions

This API is modelled on an LSP code action request. Input and output are both JSON, the request being a CodeActionParams object and the response being a list of CodeActionOrCommand objects.

debug_panic

Trigger a panic inside the wasm module. This is only useful in development for testing panic handling.

format

get_config

Docs: https://prisma.github.io/prisma-engines/doc/prisma_fmt/fn.get_config.html

lint

native_types

preview_features

referential_actions

text_document_completion

The API is modelled on an LSP completion request. Input and output are both JSON, the request being a CompletionParams object and the response being a CompletionList object.

version